legal_document_intake/requirements.txt
fastapi>=0.116.1
uvicorn[standard]>=0.35.0
pydantic>=2.11.7
python-multipart>=0.0.20
python-dotenv>=1.1.1
docling>=2.41.0
openai>=1.95.1
instructor>=1.9.2
langgraph>=0.5.2
reportlab>=4.4.2
streamlit>=1.46.1
requests>=2.32.4
pandas>=2.3.1
pytest>=8.4.1
pytest-asyncio>=1.0.0
httpx>=0.27.0
